@@ -7,11 +7,11 @@
     "html2canvas": "^1.0.0-rc.5",
     "jspdf": "^1.5.3",
     "prop-types": "^15.7.2",
-    "react": "^16.10.2",
+    "react": "^16.12.0",
     "react-bootstrap": "^1.0.0-beta.10",
     "react-dom": "^16.10.2",
     "react-router-dom": "^5.0.1",
-    "react-scripts": "3.0.1",
+    "react-scripts": "^3.3.1",
     "reactstrap": "^8.0.1"
   },
   "scripts": {
@@ -34,5 +34,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "eslint-plugin-react-hooks": "^2.3.0"
   }
 }
